Improve record and spectrogram views
- Add bottom part with buttons
- Remove spectrogram internal data from IdFlow
- Add functionality to display time

+++ naturblick/models/SoundData.swift
@@ -8,8 +8,8 @@ import Foundation
 struct SoundData {
     var sound: NBSound? = nil
     var crop: NBImage? = nil
-    var start: CGFloat? = nil
-    var end: CGFloat? = nil
+    var start: Int? = nil
+    var end: Int? = nil
     var result: [SpeciesResult]? = nil
     
     var identified: Identified? {
